requirements.txt
anyio==3.6.2
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
fastapi==0.85.2
google-api-core==2.10.2
google-auth==2.14.0
google-cloud-bigquery==3.3.5
google-cloud-bigquery-storage==2.16.2
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.50.0
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
numpy==1.23.4
packaging==21.3
proto-plus==1.22.1
protobuf==4.21.9
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.0
websockets==10.4
